RFC: Database to store download state & metadata (#116)
* Add sqlite database support

* Add state storage for content urls

* Add state storage for post and post_content

* Handle pure text posts (converted_at = null)

* Fix redirection url filename

* Do URL cache possibility check in perform_download instead of relying on use_db parameter

* Fix broken filename for non-S3 (e.g. blog images) URLs

* Make some changes

* Check for DB connection when performing updates

* Update README

* Check for post date when verifying download complete

* Add --db-bypass-post-check flag

* Update post converted_at stored in db

* Remove debug

* Update arg, README

* Update arg to mention creation

---------

+++ models.py
@@ -6,6 +6,7 @@ from requests.adapters import HTTPAdapter, Retry
 import requests
 
 from datetime import datetime as dt
+from email.utils import parsedate_to_datetime
 from urllib.parse import unquote
 from urllib.parse import urljoin
 from urllib.parse import urlparse
@@ -19,7 +20,7 @@ import sys
 import time
 import traceback
 
-
+from db import FantiaDlDatabase
 import fantiadl
 
 FANTIA_URL_RE = re.compile(r"(?:https?://(?:(?:www\.)?(?:fantia\.jp/(fanclubs|posts)/)))([0-9]+)")
@@ -66,7 +67,7 @@ class FantiaClub:
 
 
 class FantiaDownloader:
-    def __init__(self, session_arg, chunk_size=1024 * 1024 * 5, dump_metadata=False, parse_for_external_links=False, download_thumb=False, directory=None, quiet=True, continue_on_error=False, use_server_filenames=False, mark_incomplete_posts=False, month_limit=None, exclude_file=None):
+    def __init__(self, session_arg, chunk_size=1024 * 1024 * 5, dump_metadata=False, parse_for_external_links=False, download_thumb=False, directory=None, quiet=True, continue_on_error=False, use_server_filenames=False, mark_incomplete_posts=False, month_limit=None, exclude_file=None, db_path=None, db_bypass_post_check=False):
         # self.email = email
         # self.password = password
         self.session_arg = session_arg
@@ -82,6 +83,9 @@ class FantiaDownloader:
         self.month_limit = dt.strptime(month_limit, "%Y-%m") if month_limit else None
         self.exclude_file = exclude_file
         self.exclusions = []
+        self.db = FantiaDlDatabase(db_path)
+        self.db_bypass_post_check = db_bypass_post_check
+
         self.initialize_session()
         self.login()
         self.create_exclusions()
@@ -350,17 +354,9 @@ class FantiaDownloader:
             else:
                 page_number += 1
 
-    def perform_download(self, url, filepath, use_server_filename=False):
+    def perform_download(self, url, filepath, use_server_filename=False, append_server_extension=False):
         """Perform a download for the specified URL while showing progress."""
-        request = self.session.get(url, stream=True)
-
-        if request.status_code == 404:
-            self.output("Download URL returned 404. Skipping...\n")
-            return
-
-        request.raise_for_status()
-
-        url_path = unquote(request.url.split("?", 1)[0])
+        url_path = unquote(url.split("?", 1)[0])
         server_filename = os.path.basename(url_path)
         filename = os.path.basename(filepath)
         if use_server_filename:
@@ -374,14 +370,40 @@ class FantiaDownloader:
             self.output("Filename in exclusion list (skipping): {}\n".format(filename))
             return
 
+        if self.db.conn and self.db.is_url_downloaded(url_path):
+            self.output("URL already downloaded. Skipping...\n")
+            return
+
+        request = self.session.get(url, stream=True)
+        if request.status_code == 404:
+            self.output("Download URL returned 404. Skipping...\n")
+            return
+        request.raise_for_status()
+
+        # Handle redirects so we can properly catch an excluded filename
+        # Attachments typically route from fantia.jp/posts/#/download/#
+        # Images typically are served directly from cc.fantia.jp
+        # Metadata images typically are served from c.fantia.jp
+        if request.url != url:
+            url_path = unquote(request.url.split("?", 1)[0])
+            server_filename = os.path.basename(url_path)
+            if server_filename in self.exclusions:
+                self.output("Server filename in exclusion list (skipping): {}\n".format(server_filename))
+                return
+            if use_server_filename:
+                filepath = os.path.join(os.path.dirname(filepath), server_filename)
+
+        if not use_server_filename and append_server_extension:
+            filepath += guess_extension(request.headers["Content-Type"], url)
+
         file_size = int(request.headers["Content-Length"])
         if os.path.isfile(filepath) and os.stat(filepath).st_size == file_size:
             self.output("File found (skipping): {}\n".format(filepath))
+            self.db.insert_url(url_path)
             return
 
         self.output("File: {}\n".format(filepath))
-        base_filename, original_extension = os.path.splitext(filepath)
-        incomplete_filename = base_filename + ".incomplete"
+        incomplete_filename = filepath + ".part"
 
         downloaded = 0
         with open(incomplete_filename, "wb") as file:
@@ -392,9 +414,15 @@ class FantiaDownloader:
                 percent = int(100 * downloaded / file_size)
                 self.output("\r|{0}{1}| {2}% ".format("\u2588" * done, " " * (25 - done), percent))
         self.output("\n")
+
+        if downloaded != file_size:
+            raise Exception("Downloaded file size mismatch (expected {}, got {})".format(file_size, downloaded))
+
         if os.path.exists(filepath):
             os.remove(filepath)
         os.rename(incomplete_filename, filepath)
+
+        self.db.insert_url(url_path)
 
         modification_time_string = request.headers["Last-Modified"]
         modification_time = int(dt.strptime(modification_time_string, "%a, %d %b %Y %H:%M:%S %Z").timestamp())
@@ -414,8 +442,18 @@ class FantiaDownloader:
 
     def download_post_content(self, post_json, post_directory, post_title):
         """Parse the post's content to determine whether to save the content as a photo gallery or file."""
-        if post_json.get("visible_status") == "visible":
-            if post_json.get("category") == "photo_gallery":
+        self.output(f"> Content {post_json['id']}\n")
+
+        if self.db.conn and self.db.is_post_content_downloaded(post_json["id"]):
+            self.output("Post content already downloaded. Skipping...\n")
+            return True
+
+        if post_json["visible_status"] != "visible":
+            self.output("Post content not available on current plan. Skipping...\n")
+            return False
+
+        if post_json.get("category"):
+            if post_json["category"] == "photo_gallery":
                 photo_gallery = post_json["post_content_photos"]
                 photo_counter = 0
                 gallery_directory = os.path.join(post_directory, sanitize_for_path(post_title))
@@ -424,17 +462,17 @@ class FantiaDownloader:
                     photo_url = photo["url"]["original"]
                     self.download_photo(photo_url, photo_counter, gallery_directory)
                     photo_counter += 1
-            elif post_json.get("category") == "file":
+            elif post_json["category"] == "file":
                 filename = os.path.join(post_directory, post_json["filename"])
                 download_url = urljoin(POSTS_URL, post_json["download_uri"])
                 self.download_file(download_url, filename, post_directory)
-            elif post_json.get("category") == "embed":
+            elif post_json["category"] == "embed":
                 if self.parse_for_external_links:
                     # TODO: Check what URLs are allowed as embeds
                     link_as_list = [post_json["embed_url"]]
                     self.output("Adding embedded link {0} to {1}.\n".format(post_json["embed_url"], CRAWLJOB_FILENAME))
                     build_crawljob(link_as_list, self.directory, post_directory)
-            elif post_json.get("category") == "blog":
+            elif post_json["category"] == "blog":
                 blog_comment = post_json["comment"]
                 blog_json = json.loads(blog_comment)
                 photo_counter = 0
@@ -447,12 +485,15 @@ class FantiaDownloader:
                         photo_counter += 1
             else:
                 self.output("Post content category \"{}\" is not supported. Skipping...\n".format(post_json.get("category")))
+                return False
 
-            if self.parse_for_external_links:
-                post_description = post_json["comment"] or ""
-                self.parse_external_links(post_description, os.path.abspath(post_directory))
-        else:
-            self.output("Post content not available on current plan. Skipping...\n")
+        self.db.insert_post_content(post_json["id"], post_json["parent_post"]["url"].rsplit("/", 1)[1], post_json["title"], post_json["category"], post_json["foreign_plan_price"], post_json["currency_code"])
+
+        if self.parse_for_external_links:
+            post_description = post_json["comment"] or ""
+            self.parse_external_links(post_description, os.path.abspath(post_directory))
+
+        return True
 
     def download_thumbnail(self, thumb_url, post_directory):
         """Download a thumbnail to the post's directory."""
@@ -462,6 +503,11 @@ class FantiaDownloader:
 
     def download_post(self, post_id):
         """Download a post to its own directory."""
+        db_post = self.db.find_post(post_id)
+        if self.db_bypass_post_check and self.db.conn and db_post and db_post["download_complete"]:
+            self.output("Post {} already downloaded. Skipping...\n".format(post_id))
+            return
+
         self.output("Downloading post {}...\n".format(post_id))
 
         post_html_response = self.session.get(POST_URL.format(post_id))
@@ -481,6 +527,21 @@ class FantiaDownloader:
         post_title = post_json["title"]
         post_contents = post_json["post_contents"]
 
+        post_posted_at = int(parsedate_to_datetime(post_json["posted_at"]).timestamp())
+        post_converted_at = int(dt.fromisoformat(post_json["converted_at"]).timestamp()) if post_json["converted_at"] else post_posted_at
+
+        if self.db.conn and db_post and db_post["download_complete"]:
+            # Check if the post date changed, which may indicate new contents were added
+            if db_post["converted_at"] != post_converted_at:
+                self.output("Post date does not match date in database. Checking for new contents...\n")
+                self.db.update_post_download_complete(post_id, download_complete=0)
+                self.db.update_post_converted_at(post_id, post_converted_at)
+            else:
+                self.output("Post appears to have been downloaded completely. Skipping...\n".format(post_id))
+                return
+        if self.db.conn and not db_post:
+            self.db.insert_post(post_id, post_title, post_json["fanclub"]["id"], post_posted_at, post_converted_at)
+
         post_directory_title = sanitize_for_path(str(post_id))
 
         post_directory = os.path.join(self.directory, sanitize_for_path(post_creator), post_directory_title)
@@ -498,9 +559,16 @@ class FantiaDownloader:
             # Main post
             post_description = post_json["comment"] or ""
             self.parse_external_links(post_description, os.path.abspath(post_directory))
+
+        download_complete_counter = 0
         for post_index, post in enumerate(post_contents):
             post_title = post_titles[post_index]
-            self.download_post_content(post, post_directory, post_title)
+            if self.download_post_content(post, post_directory, post_title):
+                download_complete_counter += 1
+        if self.db.conn and download_complete_counter == len(post_contents):
+            self.output("All post content appears to have been downloaded. Marking as complete in database...\n")
+            self.db.update_post_download_complete(post_id)
+
         if not os.listdir(post_directory):
             self.output("No content downloaded for post {}. Deleting directory.\n".format(post_id))
             os.rmdir(post_directory)
